Final changes
* optimizations in local search (upper bound)
* heuristic classes (IteratedLocalSearch and Genetic) return status
* search result in .csv format (using the recently added csvlib)
* added results of both heuristics to data/report.xlsx

+++ mlp/include/iparser/defines.h
@@ -1,4 +1,4 @@
 #pragma
 using Node = std::size_t;
-using Pos = double;
+using Pos = float;
 using Dist = int;
